requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
matplotlib==3.5.0
numpy==1.21.4
omegaconf==2.1.1
optuna==2.10.0
Pillow==9.0.0
scikit_learn==1.0.2
torch==1.10.1
torchvision==0.11.2
transformers==4.15.0
